Add new features

+++ src/SDK/XFiler.SDK/Services/Icons/IIconLoader.cs
@@ -4,6 +4,6 @@ namespace XFiler.SDK
 {
     public interface IIconLoader
     {
-        ImageSource? GetIcon(XFilerRoute? route, double size);
+        ImageSource? GetIcon(XFilerRoute? route, int size);
     }
 }
